Add object contours to a Frames object Creates a Frames object containing raw information, combined with the segmented images and the relative trajectory under analysis

If a TrajectorySet is provided and mode is set to trajectories, returns a Frames with all trajectories included in the IDs vector painted accordingly. If the mode is set to particles, it will just plot the particles (all) on all frames. If no TrajectorySet is provided, it will be computed with default parameters. If no binary.frames is provided, it will be computed also with default parameters

add.contours(
  raw.frames,
  binary.frames = NULL,
  trajectoryset = NULL,
  trajIDs = NULL,
  mode = "particles",
  col = NULL,
  channel = NULL
)

Arguments

raw.frames
A Frames object with raw images

binary.frames
A Frames object with preprocessed frames

trajectoryset
A TrajectorySet object

trajIDs
Numeric vector, the ID(s) of the trajectory.

mode
A character string, can assume the values particles or trajectories. Defaults to particles

col
A vector of color strings

channel
A character string, to select which channel to process

Value

A new Frames object with contours of the objects added

Examples

data("MesenteriumSubset")
if (FALSE) {
paintedTrajectories <- add.contours(raw.frames = MesenteriumSubset,
                                    mode = "trajectories",channel="red")
paintedParticles <- add.contours(raw.frames = MesenteriumSubset,
                                 mode = "particles",channel="red")
inspect.Frames(paintedTrajectories)
inspect.Frames(paintedParticles)
}