Channel extraction for objects

channel

channel.Frames(frames, mode)

Arguments

frames

A Frames object

mode

A character value specifying the target mode for conversion.

Value

A Frames object with just the infotmation on the selected channel

Examples

data("MesenteriumSubset")
channel.Frames(MesenteriumSubset,"red")
#> Frames 
#>   colorMode    : Grayscale 
#>   storage.mode : double 
#>   dim          : 271 131 20 
#>   frames.total : 20 
#>   frames.render: 20 
#> 
#> imageData(object)[1:5,1:6,1]
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.1647059 0.2117647 0.1882353 0.1803922 0.1607843 0.1333333
#> [2,] 0.2352941 0.1882353 0.1803922 0.1568627 0.1411765 0.1372549
#> [3,] 0.2352941 0.2000000 0.1764706 0.1490196 0.1333333 0.1333333
#> [4,] 0.2352941 0.2117647 0.1764706 0.1529412 0.1411765 0.1411765
#> [5,] 0.2313725 0.2078431 0.1725490 0.1411765 0.1294118 0.1411765