Cut borders of a Frames object

Performs cropping on the Frames object, selecting how many pixels should be cut on each side

crop.Frames(
  frames,
  cutLeft = 5,
  cutRight = 5,
  cutUp = 5,
  cutDown = 5,
  cutAll = 0,
  testing = FALSE,
  ...
)

Arguments

frames

An input Frames object

cutLeft

Amount of pixels to be cut at the side

cutRight

Amount of pixels to be cut at the side

cutUp

Amount of pixels to be cut at the side

cutDown

Amount of pixels to be cut at the side

cutAll

Amount of pixels to be cut at all sides. Overrides the single side values

testing

Logical, whether to just test the cropping or to actually perform it. Default set to FALSE

...

Arguments to be passed to display (e.g. setting the method argument)

Value

A Frames object, with cropped frames in the image slot

Details

Cropping can be performed with careful choice of all cutting sides, or cropping a single value from all sides

Author

Federico Marini, marinif@uni-mainz.de, 2014

Examples

data("MesenteriumSubset")
crop.Frames(MesenteriumSubset)
#> Frames 
#>   colorMode    : Color 
#>   storage.mode : double 
#>   dim          : 262 122 3 20 
#>   frames.total : 60 
#>   frames.render: 20 
#> 
#> imageData(object)[1:5,1:6,1,1]
#>           [,1]      [,2]      [,3]      [,4]      [,5]       [,6]
#> [1,] 0.1294118 0.1411765 0.1372549 0.1294118 0.1019608 0.09411765
#> [2,] 0.1372549 0.1372549 0.1333333 0.1176471 0.1098039 0.09019608
#> [3,] 0.1490196 0.1254902 0.1215686 0.1215686 0.1254902 0.11372549
#> [4,] 0.1568627 0.1294118 0.1176471 0.1019608 0.1411765 0.12549020
#> [5,] 0.1529412 0.1333333 0.1254902 0.1333333 0.1490196 0.13725490