Exports a Frames object

Writes the images contained in the image slot of the Frames object elements. The images can be exported as single frames, or as a .gif image that is composed by the single frames.

export.Frames(frames, dir = tempdir(), nameStub = "testExport",
  createGif = FALSE, removeAfterCreatingGif = TRUE)

Arguments

frames
A Frames object

dir
The path of the folder where the image should be written

nameStub
The stub for the file name, that will be used as a prefix for the exported images

createGif
Logical, whether to create or not an animated .gif file

removeAfterCreatingGif
Logical, whether to remove the single exported .png images after creating the single .gif

Value

Image files are written in the desired location

Examples

data("MesenteriumSubset")
# NOT RUN {
export.Frames(MesenteriumSubset,nameStub="subset_export_",
                       createGif=TRUE,removeAfterCreatingGif=FALSE)
# }