Links a ParticleSet object

Performs linking of the particles by tracking them through the frames

link.particles(particleset, L, R = 2, epsilon1 = 0.1, epsilon2 = 2,
  lambda1 = 1, lambda2 = 1,
  penaltyFunction = penaltyFunctionGenerator(), verboseOutput = FALSE,
  prog = FALSE, include.intensity = TRUE, include.area = FALSE)

Arguments

particleset
A ParticleSet object

L
Maximum number of pixels an object can move in two consecutive frames

R
Linkrange, i.e. the number of consecutive frames to search for potential candidate links

epsilon1
A numeric value, to be used in the formula. Jitter for allowing angular displacements

epsilon2
A numeric value, to be used in the formula. Jitter for allowing spatial displacements

lambda1
A numeric value. Multiplicative factor for the penalty function

lambda2
A numeric value. Multiplicative factor applied to the angular displacement

penaltyFunction
A function structured in such a way to be applied as penalty function in the linking

verboseOutput
Logical, whether the output should report additional intermediate steps. For debugging use mainly

prog
Logical, whether the a progress bar should be shown during the tracking phase

include.intensity
Logical, whether to include also intensity change of the particles in the cost function calculation

include.area
Logical, whether to include also area change of the particles in the cost function calculation

Value

A LinkedParticleSet object

References

I F Sbalzarini and P Koumoutsakos."Feature point tracking and trajectory analysis for video imaging in cell biology." In: Journal of structural biology 151.2 (Aug. 2005), pp. 182-95. ISSN: 1047-8477. DOI: 10.1016/j.jsb.2005.06.002. URL: http://www.ncbi.nlm.nih.gov/pubmed/16043363

Examples

data("candidate.platelets")
tracked.platelets <- link.particles(candidate.platelets, L= 40)