Match trajectories to related particles.

Match trajectories to the related particles in the TrajectorySet and ParticleSet objects. This function returns a new ParticleSet object that contains as additional column the trajectory ID that the particular particle was assigned to. Used also by other routines, such as snap

matchTrajToParticles(particleset, trajectoryset)

Arguments

particleset
A ParticleSet object

trajectoryset
A TrajectorySet object coupled to the particleset

Value

A ParticleSet object with an additional column with the trajectory IDs

Examples

data(candidate.platelets)
trajs <- trajectories(candidate.platelets)

#> Input ParticleSet is not a LinkedParticleSet.
#> Performing linking first with some set of default parameters - you might want to change them according to your scenario...

matchTrajToParticles(candidate.platelets, trajs)

#> An object of the ParticleSet class. 
#> 
#> Set of particles for 20 images
#> 
#> Displaying a subset of the features of the 14 particles found in the first image...
#>   cell.0.m.cx cell.0.m.cy cell.0.m.majoraxis cell.0.m.eccentricity
#> 1   186.70833   47.937500           8.916405             0.6353715
#> 2   256.19048   35.857143           7.746554             0.4606665
#> 3   251.09524   63.523810           8.552466             0.6843186
#> 4    15.79412    8.058824           8.419655             0.7892764
#> 5   215.54688   51.828125          13.694783             0.8788344
#>   cell.0.m.theta cell.0.s.area cell.0.s.perimeter cell.0.s.radius.mean
#> 1      0.3380463            48                 20             3.487042
#> 2      0.9165252            42                 19             3.194559
#> 3      0.9370618            42                 19             3.247531
#> 4      1.4806734            34                 18             2.847459
#> 5      0.7172299            64                 28             4.308530
#> 
#> Particles identified on the red channel