Extracts particles from the images of a Frames object.

Extracts particles from the images of a Frames object.

particles(raw.frames, binary.frames = NULL, channel = NULL,
  BPPARAM = bpparam())

Arguments

raw.frames
A Frames object with the raw images (mandatory)

binary.frames
A Frames object with preprocessed images (optional, if not provided gets produced with standard default parameters)

channel
Character string. The channel to perform the operations on. Can be red, green or blue

BPPARAM
a MulticoreParam object, used to control the performances inside the BiocParallel call to process frames in parallel by taking advantage of the computing infrastructure available

Value

A ParticleSet object, containing all detected particles for each frame

Examples

data("MesenteriumSubset")