Constructor for a Frames object

This function is used to create a Frames object from a vector of image files (or a folder specifying the directory containing them). The number of frames is also specified, as just a subset of the images can be used for this

read.Frames(image.files, nframes = NULL)

Arguments

image.files
Vector of strings containing the locations where the (raw) images are to be found, or alternatively, the path to the folder

nframes
Number of frames that will constitute the Frames object

Value

An object of the Frames class, which holds the info on a list of frames, specifying for each the following elements:

image

The Image object containing the image itself

location

The complete path to the location of the original image

Examples

## see vignette

# NOT RUN {
fullData <- read.Frames(image.files = "/path/to/the/directory", nframes = 100)
# }