Display conveniently a TrajectorySet object

Display conveniently a TrajectorySet object

# S4 method for TrajectorySet
show(object)

Arguments

object

A TrajectorySet object

Value

This returns an invisible NULL.

Author

Federico Marini, marinif@uni-mainz.de, 2014

Examples

data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
#> Input ParticleSet is not a LinkedParticleSet.
#> Performing linking first with some set of default parameters - you might want to change them according to your scenario...
print(platelets.trajectories)
#> An object of the TrajectorySet class. 
#> 
#> TrajectorySet composed of 25 trajectories
#> 
#> Trajectories cover a range of 20 frames
#> Displaying a segment of the first trajectory...
#>        xCoord   yCoord trajLabel frame frameobjectID
#> 1_1  186.7083 47.93750         1     1             1
#> 1_2  186.9649 48.26316         1     2             4
#> 1_3  186.8136 48.18644         1     3             2
#> 1_4  186.2807 47.70175         1     4             1
#> 1_5  186.6897 47.87931         1     5             2
#> 1_6  186.8269 48.11538         1     6             2
#> 1_7  186.9643 48.30357         1     7             1
#> 1_8  186.6207 48.36207         1     8             3
#> 1_9  186.3273 48.05455         1     9             3
#> 1_10 186.9821 48.19643         1    10             3
#> 
#> Trajectories are related to particles identified on the red channel