Snap the features of the closest particle identified

This function combines all classes related to a single experiment in order to deliver a clickable feedback on one of the frames.

snap(raw.frames, binary.frames, particleset, trajectoryset, frameID = 1,
  infocol = "yellow", infocex = 1, showVelocity = FALSE)

Arguments

raw.frames
A Frames object with the raw frames data

binary.frames
A Frames object with the preprocessed frames data

particleset
A ParticleSet object with the particles data

trajectoryset
A TrajectorySet object with the trajectories data

frameID
The ID of the frame to inspect

infocol
The color to use for plotting the contours and the information on the clicked particle

infocex
The numeric character expansion value as in cex to be used for printing the text on the image

showVelocity
Logical, whether to display additional information on the instantaneous velocity of the particle

Value

An image of the selected frame, rendered in R native graphics, and additionally a list with the coordinates as well as the trajectory ID of the particle closest to the clicked location

Examples

# NOT RUN {
data(MesenteriumSubset)
binary.frames <- preprocess.Frames(channel.Frames(MesenteriumSubset,"red"))
particleset <- particles(MesenteriumSubset,binary.frames,"red")
trajectoryset <- trajectories(particleset)
snap(MesenteriumSubset,binary.frames,particleset,trajectoryset,frameID=1)
# }