Calculate a set of kinematics parameter from a TrajectorySet object, or a single parameter, or from a single trajectory (all possible combinations)

The computed set of parameters include delta.x, delta.t and delta.v (displacements and instantaneous velocity), totalTime, totalDistance, distStartToEnd, curvilinearVelocity, straightLineVelocity and linearityForwardProgression, Mean Squared Displacement, velocity autocorrelation, and more. If a single trajectory is specified, the computation is performed for that trajectory alone. If a parameter is specified, only that parameter is reported, either for one or all trajectories

kinematics(
  trajectoryset,
  trajectoryIDs = NULL,
  acquisitionFrequency = 30,
  scala = 50,
  feature = NULL
)

Arguments

trajectoryset

A TrajectorySet object

trajectoryIDs

The ID of a single trajectory

acquisitionFrequency

The frame rate of acquisition for the images, in milliseconds

scala

The value of micro(?)meters to which each single pixel corresponds

feature

Character string, the name of the feature to be computed

Value

A KinematicsFeaturesSet object, or a KinematicsFeatures object, or an atomic value, or a list(eventually coerced to a vector)

Author

Federico Marini, marinif@uni-mainz.de, 2014

Examples

data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
#> Input ParticleSet is not a LinkedParticleSet.
#> Performing linking first with some set of default parameters - you might want to change them according to your scenario...
# for all trajectories, all features
alltrajs.features <- kinematics(platelets.trajectories)
#> Warning: The trajectory with ID 17 had 3 or less points, no features were computed.
#> Warning: The trajectory with ID 19 had 3 or less points, no features were computed.
#> Warning: The trajectory with ID 21 had 3 or less points, no features were computed.
#> Warning: The trajectory with ID 25 had 3 or less points, no features were computed.
# for one trajectory, all features
traj11features <- kinematics(platelets.trajectories,trajectoryIDs = 11)
# for all trajectories, one feature
alltrajs.curvVel <- kinematics(platelets.trajectories,feature = "curvilinearVelocity")
#> Warning: The trajectory with ID 17 had 3 or less points, no features were computed.
#> Warning: The trajectory with ID 19 had 3 or less points, no features were computed.
#> Warning: The trajectory with ID 21 had 3 or less points, no features were computed.
#> Warning: The trajectory with ID 25 had 3 or less points, no features were computed.