Preprocessing function for Frames objects

Frames objects are processed according to the chosen set of parameters. Many of them refer directly to existing EBImage functions, please see the corresponding help for additional information

preprocess.Frames(
  frames,
  brush.size = 3,
  brush.shape = "disc",
  at.offset = 0.15,
  at.wwidth = 10,
  at.wheight = 10,
  kern.size = 3,
  kern.shape = "disc",
  ws.tolerance = 1,
  ws.radius = 1,
  displayprocessing = FALSE,
  ...
)

Arguments

frames
A Frames object

brush.size
Size in pixels of the brush to be used for initial smoothing (low-pass filtering)

brush.shape
Shape of the brush to be used for initial smoothing (low-pass filtering)

at.offset
Offset to be used in the adaptive thresholding step - see also thresh. As an alternative thresholding method, see also otsu in the EBImage package.

at.wwidth
Width of the window for the adaptive thresholding step - see also thresh. As an alternative thresholding method, see also otsu in the EBImage package.

at.wheight
Height of the window for the adaptive thresholding step - see also thresh. As an alternative thresholding method, see also otsu in the EBImage package.

kern.size
Size in pixels of the kernel used for morphological operations - e.g., opening, which is an erosion followed by a dilation, and closing which is a dilation followed by an erosion - see also opening, closing

kern.shape
Shape of the kernel used for morphological operations

ws.tolerance
Tolerance allowed in performing the watershed-based segmentation (see also watershed)

ws.radius
Radius for the watershed-based segmentation (see also watershed)

displayprocessing
Logical, whether to display intermediate steps while performing preprocessing. Dismissed currently, it could increase runtime a lot

...
Arguments to be passed to methods

Value

A Frames object, whose frame images are the preprocessed versions of the input images

Author

Federico Marini, marinif@uni-mainz.de, 2014

Examples

data("MesenteriumSubset")
preprocess.Frames(channel.Frames(MesenteriumSubset,"red"))

#> Frames 
#>   colorMode    : Grayscale 
#>   storage.mode : integer 
#>   dim          : 271 131 20 
#>   frames.total : 20 
#>   frames.render: 20 
#> 
#> imageData(object)[1:5,1:6,1]
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    0    0    0    0    0    0
#> [2,]    0    0    0    0    0    0
#> [3,]    0    0    0    0    0    0
#> [4,]    0    0    0    0    0    0
#> [5,]    0    0    0    0    0    0
#> 
#> Channel(s): red