Display conveniently a KinematicsFeatureSet object

Display conveniently a KinematicsFeatureSet object

# S4 method for KinematicsFeaturesSet
show(object)

Arguments

object
A KinematicsFeatureSet object

...
Arguments to be passed to methods

Value

This returns an invisible NULL.

Examples

data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)

#> Input ParticleSet is not a LinkedParticleSet.
#> Performing linking first with some set of default parameters - you might want to change them according to your scenario...

alltrajs.features <- kinematics(platelets.trajectories)

#> Warning: The trajectory with ID 17 had 3 or less points, no features were computed.

#> Warning: The trajectory with ID 19 had 3 or less points, no features were computed.

#> Warning: The trajectory with ID 21 had 3 or less points, no features were computed.

#> Warning: The trajectory with ID 25 had 3 or less points, no features were computed.

print(alltrajs.features)

#> An object of the KinematicsFeaturesSet class. 
#> 
#> KinematicsFeaturesSet composed of 25  KinematicsFeatures objects
#> 
#> Available features (shown for the first trajectory):
#>  [1] "delta.x"                     "delta.t"                    
#>  [3] "delta.v"                     "totalTime"                  
#>  [5] "totalDistance"               "distStartToEnd"             
#>  [7] "curvilinearVelocity"         "straightLineVelocity"       
#>  [9] "linearityForwardProgression" "trajMSD"                    
#> [11] "velocityAutoCorr"            "instAngle"                  
#> [13] "directChange"                "dirAutoCorr"                
#> [15] "paramsNotComputed"          
#> 
#> Curvilinear Velocity: 0.009970094 
#> Total Distance: 5.682953 
#> Total Time: 570 
#> 
#> Average values (calculated on 4 trajectories where parameters were computed)
#> Average Curvilinear Velocity: 0.1146995 
#> Average Total Distance: 48.49261 
#> Average Total Time: 484.2857